WIP: Reorganize project structure and create master pipeline
- Reorganized project with proper subdirectories (tests/, scripts/, examples/, data/)
- Fixed mixed precision issues by defaulting to float32 instead of bfloat16
- Fixed data processor to write binary files to correct output directory
- Created comprehensive master pipeline script (run_pipeline.py) that:
  - Cleans old data by removing ./data directory entirely
  - Downloads and processes fresh training data
  - Trains model with specified configuration
  - Runs test suite
  - Creates examples (placeholder)
- Updated all configurations to use proper data paths and float32
- Fixed trainer data loader initialization and mixed precision API usage
- Moved train.py, generate.py to scripts/ directory
- Moved test_*.py files to tests/ directory
- Updated configs to point to data/ directory for binary files

Ready to run full pipeline for clean training from scratch.

+++ Gemma-270M/gemma_270m/config.py
@@ -46,7 +46,7 @@ class ModelConfig:
     ])
     
     # Data type and attention scaling
-    dtype: torch.dtype = torch.bfloat16
+    dtype: torch.dtype = torch.float32
     query_pre_attn_scalar: int = 256
     
     def __post_init__(self):
@@ -94,7 +94,7 @@ class TrainingConfig:
     grad_clip: float = 1.0
     
     # Mixed precision and device
-    dtype: str = "bfloat16"  # "float32", "bfloat16", "float16" 
+    dtype: str = "float32"  # "float32", "bfloat16", "float16"
     device: str = "cuda"
     compile_model: bool = False
